+++ Runtime/StateGraph/States/BaseParallelSubStateMachineState.cs
@@ -78,7 +78,7 @@ namespace Nonatomic.VSM2.StateGraph.States
 			{
 				if (model == null) continue;
 				
-				var subSubStateMachine = new StateMachine(model, this.GameObject);
+				var subSubStateMachine = new StateMachine(model, GameObject, SharedData);
 				subSubStateMachine.SetParent(StateMachine);
 				SubStateMachines.Add(subSubStateMachine);
 			}